+++ src/refactoring/entities/product/index.ts
@@ -1,1 +1,2 @@
 export * from "./model";
+export * from "./ui";
