+++ src/refactoring/entities/product/model/discount.ts
@@ -2,3 +2,7 @@ export interface Discount {
   quantity: number;
   rate: number;
 }
+
+export const getMaxDiscount = (discounts: Discount[]) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
