+++ src/refactoring/entities/product/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductInfoCard } from "./ProductInfoCard";
