+++ src/refactoring/features/cart/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as AddToCartButton } from "./AddToCartButton";
